Agregar exceptions, security fase 1, test terminados

+++ src/main/java/com/example/finalProject/Services/ClienteService.java
@@ -4,14 +4,13 @@ import com.example.finalProject.Exception.InvalidStatementException;
 import com.example.finalProject.Modules.Cliente;
 import com.example.finalProject.Repositories.ClienteRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.util.Streamable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 @Service
 public class ClienteService {
@@ -29,17 +28,21 @@ public class ClienteService {
     }
 
 
-    public ResponseEntity crearCliente(Cliente cliente) {
+    public Cliente crearCliente(Cliente cliente) {
         if (verificarClienteExiste(cliente.getCedula())){
-            return new ResponseEntity("Ya existe un cliente con esta cédula", HttpStatus.BAD_REQUEST);
+            throw new InvalidStatementException("Ya existe un cliente con la cédula "+cliente.getCedula() + " en el sistema");
         }
         else{
-            if((cliente.getCedula().toString()).length() != 10){
-            this.clienteRepository.save(cliente);
-            return new ResponseEntity(cliente,HttpStatus.ACCEPTED);
+            if((cliente.getCedula().toString()).length() == 10){
+                if(validateEmail(cliente.getCorreoElectronico())){
+                    return this.clienteRepository.save(cliente);
+                }
+                else{
+                    throw new InvalidStatementException("El correo electrónico no es válido");
+                }
             }
             else{
-                return new ResponseEntity("La cédula debe tener 10 dígitos", HttpStatus.BAD_REQUEST);
+                throw new InvalidStatementException("La cédula debe tener 10 dígitos");
             }
         }
     }
@@ -48,37 +51,41 @@ public class ClienteService {
         return this.clienteRepository.findAll();
     }
 
-    public ResponseEntity obtenerClienteCedula(Integer cedula) {
-        try{
-            return ResponseEntity.ok(this.clienteRepository.findById(cedula));
+    public Cliente obtenerClienteCedula(Integer cedula) {
+        if (verificarClienteExiste(cedula)){
+            return this.clienteRepository.findById(cedula).get();
         }
-        catch (Exception e){
-            return new ResponseEntity("No existe un cliente con esta cédula", HttpStatus.BAD_REQUEST);
-        }
+
+            throw new InvalidStatementException("No existe un cliente con la cédula " + cedula + " en el sistema");
     }
 
 
-    public ResponseEntity actualizarCliente (Cliente cliente) {
+    public Cliente actualizarCliente (Cliente cliente) {
         if(verificarClienteExiste(cliente.getCedula())){
-            this.clienteRepository.save(cliente);
-            return ResponseEntity.ok(obtenerClienteCedula(cliente.getCedula()));
+            return this.clienteRepository.save(cliente);
         }
         else {
-            return new ResponseEntity("No existe un cliente con esta cédula", HttpStatus.BAD_REQUEST);
+            throw new InvalidStatementException("No existe un cliente con la cédula " + cliente.getCedula() + " en el sistema");
         }
-
     }
 
 
     public ResponseEntity eliminar(Integer cedula) {
-        try{
-            this.clienteRepository.deleteById(cedula);
-            return new ResponseEntity("El cliente " + cedula+ " se ha eliminado con éxito", HttpStatus.ACCEPTED);
-
+        if (!verificarClienteExiste(cedula)){
+            throw new InvalidStatementException("No existe un cliente la cédula " + cedula + " en el sistema");
         }
-        catch (Exception e){
-            return new ResponseEntity("No existe un cliente con esta cédula", HttpStatus.BAD_REQUEST);
-            }
+        else{
+            this.clienteRepository.deleteById(cedula);
+            return new ResponseEntity("El cliente con cédula " + cedula+ " se ha eliminado con éxito", HttpStatus.ACCEPTED);
+        }
+    }
+
+    public boolean validateEmail(String email) {
+        Pattern pattern = Pattern
+                .compile("^[_A-Za-z0-9-]+(\\.[_A-Za-z0-9-]+)*@"
+                        + "[A-Za-z0-9-]+(\\.[A-Za-z0-9]+)*(\\.[A-Za-z]{2,})$");
+        Matcher matcher = pattern.matcher(email);
+        return matcher.find();
     }
 
 }
